+++ webcomponents/core/src/components/deck/deckdeckgo-deck/deckdeckgo-deck.tsx
@@ -302,10 +302,6 @@ export class DeckdeckgoDeck {
 
   @Method()
   async toggleKeyboardAssist(state: boolean) {
-    if (!document) {
-      return;
-    }
-
     if (!this.keyboard) {
       return;
     }