+++ webcomponents/core/src/components/deck/deckdeckgo-deck/deckdeckgo-deck.tsx
@@ -302,10 +302,6 @@ export class DeckdeckgoDeck {
 
   @Method()
   async toggleKeyboardAssist(state: boolean) {
-    if (!document) {
-      return;
-    }
-
     if (!this.keyboard) {
       return;
     }
@@ -1077,7 +1073,7 @@ export class DeckdeckgoDeck {
         return;
       }
 
-      slider.style.setProperty('cursor', show ? 'initial' : 'none');
+      slider.style.setProperty('cursor', show ? 'inherit' : 'none');
       this.mouseInactivity.emit(show);
 
       this.cursorHidden = !show;