+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Others
 
+- kit: v3.1.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/kit/CHANGELOG.md))
 - utils: v1.8.1 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/master/utils/utils/CHANGELOG.md))
 
 <a name="4.5.0"></a>